refactor: move auth api instance
