+++ src/shared/auth/auth-client.ts
@@ -1,7 +1,7 @@
 import { createAuthClient } from 'better-auth/client';
 import { adminClient, inferAdditionalFields } from 'better-auth/client/plugins';
 import { auth } from './auth';
-import { ac, admin, company, user } from './auth/permissions';
+import { ac, admin, company, user } from './permissions';
 
 export const authClient = createAuthClient({
   plugins: [
